requirements.txt
faiss-cpu
sentence-transformers
numpy
sqlite-utils
pyyaml